Refactor extents and disk reading (of user data)
With this refactoring we also move the disk_reading closer to the FUSE
interface.  This is important because at a higher level we can make better
decisions about how to schedule the reads.  It also makes lower layers a bit
more robust by removing I/O from them (more pure, as in FP, approach).