Run tests multiple times
This doesn't slow the test run too much, since the costly operation is setting
up the test partition (dd, mke2fs and so on).  Of course, now we also get some
more reliable results.  The downside is that this results can no longer be
compared to the old ones, which where not being run with a warmed up cache