FreeBSD fix for unaligned reads
Apparently, FreeBSD is not as forgiving as Linux or OSX when it comes to
unaligned reads of block devices.  This commit introduces one pread(2)
wrapper that does nothing on Linux but on FreeBSD performs aligned reads.
That means that for fully unaligned access, there are 3 pread(2) calls
instead of the usual 1.

Unfortunately, there is some code duplication with op_read.c.  But for
now let's make this work reasonably on FreeBSD and then let's worry
about style.  Might be a good idea to get rid of non aligned reads as
much as possible, and document the rest (by the use of a different API).

Thanks to William Grzybowski for recognizing the problem in the first
place.