Fixes for issue #6 and sparse files
The crash in issue #6 was caused by buf being incremented past the end
of its allocated space.  This only happened when a file was fragmented
enough to cause the op_read() loop to run more than twice, so it wasn't
caught by the existing tests.

This also adds support for sparse files.  A missing pblock is assumed to
be a "hole" in the file, rather than an error.