Get rid of ext4_inode_extent definition
This definition is making hard to cleanly generalizing extent_get_pblock to
any sort of other extent array.  Let's get rid of it.  At the same time we
no longer need to rely on the extent array of the inode being 4 entries.

Plus, this commit removes code while keeping readability and functionality
intact... so why the hell not.