Do the file magic check before fuse initialization
Once we enter the fuse main loop we are already on the background and is
not easy to provide feedback to the user, which might not have noticed
the wrong device name or file.  Linux even creates an mtab entry before the
process manages to die.  This commit moves the magic check before the
fuse_main call to avoid the problem altogether.