test library: fix occasional umount failures
Although I'm not able to pin down *why*, it seems that my Ubuntu 12.04
machine sometimes complains that my mountpoint is busy, although it
shouldn't be the case.  Seems to happen more often in very fast testcases.

After this commit, if the umount fails, we will sleep for 1/10 of a second
and retry.