Memory corruption with long symlinks
So we were happily writing an entire block to the buffer that the readlink
fuse operation was handling us.  That buffer is not necessarily big enough
for that.  In the particular case of #22, the buffer is 1025 bytes, while
the block size is the standard 4K.  Luckily, pread(2) calls us out and
refuses to corrupt the memory.